Refactor toom-cook 3
